icons

+++ src/index.ts
@@ -69,36 +69,73 @@ export class NodeInno {
 
 		let buildTempRoot = path.join(os.homedir(), "Desktop/node-inno");
 
-		console.log(defaultJson);
+		console.log(this.colorYellow, '\ndefault config :');
+		console.log(this.colorYellow, JSON.stringify(defaultJson, null, 4));
 
-		targetJson = require('extend')(true, defaultJson, targetJson);
+		targetJson.ui = targetJson.ui || {};
+		targetJson.app = targetJson.app || {};
+		targetJson.installDetail = targetJson.installDetail || {};
 
-		/**
-		 * dir config
-		 */
 		targetJson.app.name = targetJson.app.name || pakage.name;
 		targetJson.app.version = targetJson.app.version || pakage.version;
-		targetJson.app.exe = path.join(process.cwd(), targetJson.app.exe);
-		targetJson.installDetail.installerOutputDir = path.join(process.cwd(), targetJson.installDetail.installerOutputDir || `${pakage.name}-${pakage.version}.exe`);
-		targetJson.app.package = path.join(process.cwd(), targetJson.app.package);
+		
+		/*check for absolute path */
+		if(targetJson.app.exe && targetJson.app.exe.substr(1,1)==":"){
+			targetJson.app.exe = path.join('', targetJson.app.exe);
+		} else {
+			targetJson.app.exe = path.join(process.cwd(), targetJson.app.exe);
+		}
+		
+		/*check for absolute path */
+		if(targetJson.app.package && targetJson.app.package.substr(1,1)==":"){
+			targetJson.app.package = path.join("", targetJson.app.package);
+		} else {
+			targetJson.app.package = path.join(process.cwd(), targetJson.app.package);
+		}
+		
+		/*check for absolute path */
+		if(targetJson.installDetail.installerOutputDir && targetJson.installDetail.installerOutputDir.substr(1,1)==":"){
+			targetJson.installDetail.installerOutputDir = path.join(targetJson.installDetail.installerOutputDir);
+		} else {
+			targetJson.installDetail.installerOutputDir = path.join(process.cwd(), targetJson.installDetail.installerOutputDir || `${pakage.name}-${pakage.version}`);
+		}
 
-		console.log(targetJson);
+		targetJson.installDetail.defaultInstallDir = targetJson.installDetail.defaultInstallDir || `{pf}`;
 
-		/* copy resources to temp */
-		try{
-			fs.ensureDirSync(buildTempRoot);
-			fs.copySync(path.join(nodeInnoBase, "template/package"), buildTempRoot);
-		} catch(e){
-			console.error(this.colorRed, e);
+		if(!targetJson.app.exe){
+			console.error(this.colorRed, `× "${targetJson.app.exe}" required ×`);
 			return false;
 		}
 
-		fs.copySync(path.join(process.cwd(), pakage["node-inno-build"]), buildTempRoot);
+		if(!targetJson.app.package){
+			console.error(this.colorRed, `× "${targetJson.app.package}" required ×`);
+			return false;
+		}
+
+		if(!fs.existsSync(targetJson.app.exe)){
+			console.error(this.colorRed, `× "${targetJson.app.exe}" not exist ×`);
+			return false;
+		}
+
+		if(!fs.existsSync(targetJson.app.package.replace("*", ""))){
+			console.error(this.colorRed, `× "${targetJson.app.package}" not exist ×`);
+			return false;
+		}		
+
+		targetJson = require('extend')(true, defaultJson, targetJson);
+		/**
+		 * dir config
+		 */
+		console.log(this.colorYellow, '\n\ntarget config :');
+		console.log(this.colorYellow, JSON.stringify(targetJson, null, 4));
+
+		//fs.copySync(path.join(process.cwd(), pakage["node-inno-build"]), buildTempRoot);
 
 		if(!targetJson.buildScript){
 			/* compile script template*/
 			let template = require('art-template');
 			try {
+				/* copy default resources to temp */
 				fs.copySync(path.join(nodeInnoBase, "template/package"), buildTempRoot);
 	
 				let scripts = [
@@ -134,12 +171,35 @@ export class NodeInno {
 	
 				this.config = targetJson;
 	
+				/* copy custom resources to temp */
+				let customResource = path.join(process.cwd(), "build/inno-resource");
+				if(fs.existsSync(customResource)){
+					try{
+						let tempResource = path.join(buildTempRoot, "inno-resource");
+						fs.copySync(customResource, tempResource);
+					} catch(e){
+						console.log(e);
+					}
+				}
+
 				return true;
 			} catch (e) {
 				console.error(this.colorRed, e);
 				return false;
 			}
 		}
+
+		/* copy custom resources to temp */
+		let customResource = path.join(process.cwd(), "build/inno-resource");
+		if(fs.existsSync(customResource)){
+			try{
+				let tempResource = path.join(buildTempRoot, "inno-resource");
+				fs.copySync(customResource, tempResource);
+			} catch(e){
+				console.log(e);
+			}
+		}
+
 
 		return true;
 	}
@@ -159,15 +219,15 @@ export class NodeInno {
 		
 		/* 开始编译 */
 		buildProcess.stdout.on('data', (data: any) => {
-			//console.log(this.colorGreen, data);
+			console.log(this.colorGreen, data);
 		});
 
 		buildProcess.stderr.on('data', (data: any) => {
-			//console.log(this.colorRed, data);
+			console.log(this.colorRed, data);
 		});
 
 		buildProcess.on('close', (code: any) => {
-			//console.log(this.colorGreen, `child process exited with code ${code}`);
+			console.log(this.colorGreen, `child process exited with code ${code}`);
 		});
 
 		buildProcess.on('exit', (code: any) => {
